@@ -2,11 +2,11 @@
   "main": "dist/Main.js",
   "scripts": {
     "start": "node ./dist/Main.js",
-    "settle": "node ./tools/check-global.js && yarn install && node ./tools/setup.js && node ./tools/check-front.js && yarn build",
+    "settle": "yarn install && node ./tools/setup.js && node ./tools/check-front.js && yarn build",
     "build": "yarn run build-back && webpack && node ./tools/build-scss.js && node ./tools/lib/front-lang-loader.js !",
-    "build-back": "webpack -p --config ./tools/webpack.back.config.js",
+    "build-back": "webpack --mode production --config ./tools/webpack.back.config.js",
     "build-front": "node ./tools/front.js !",
-    "watch-back": "webpack -d -w --config ./tools/webpack.back.config.js",
+    "watch-back": "webpack --mode development -w --config ./tools/webpack.back.config.js",
     "watch-front": "node ./tools/front.js"
   },
   "globalDependencies": [
@@ -31,6 +31,7 @@
     "typeorm": "^0.2.18"
   },
   "devDependencies": {
+    "@tsconfig/node14": "^1.0.0",
     "@types/accept-language-parser": "^1.4.0",
     "@types/cookie-parser": "^1.4.1",
     "@types/express": "^4.11.1",
@@ -41,10 +42,13 @@
     "mysql": "^2.16.0",
     "node-sass": "^4.8.3",
     "readline": "^1.3.0",
-    "ts-json-schema-generator": "^0.58.1",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.3",
-    "webpack-node-externals": "^1.7.2"
+    "sass": "^1.29.0",
+    "ts-json-schema-generator": "^0.60.0",
+    "ts-loader": "^8.0.11",
+    "typescript": "^4.1.2",
+    "webpack": "^5.9.0",
+    "webpack-cli": "^4.2.0",
+    "webpack-node-externals": "^2.5.2"
   },
   "engines": {
     "yarn": ">= 1.0.0"
