@@ -1,7 +1,7 @@
 /**
  * `ACTION_RECEIVER_TABLE`에서 정의하는 핸들러 함수의 자료형.
  */
-declare type Action<T extends any[] = []> = (...args:T) => boolean|void;
+declare type Action<T extends any[] = []> = (...args:T) => (boolean|void);
 /**
  * 주어진 함수의 매개 변수의 자료형.
  */
